+++ main.go
@@ -1,23 +1,32 @@
 //go:generate go install -v github.com/kevinburke/go-bindata/go-bindata
-//go:generate go-bindata -prefix res/ -pkg assets -o assets/assets.go res/Supermium.lnk
+//go:generate go-bindata -prefix res/ -pkg assets -o assets/assets.go res/Firefox.lnk
 //go:generate go install -v github.com/josephspurrier/goversioninfo/cmd/goversioninfo
 //go:generate goversioninfo -icon=res/papp.ico -manifest=res/papp.manifest
 package main
 
 import (
+	"fmt"
 	"os"
 	"path"
+	"strings"
+	"text/template"
 
-	"github.com/portapps/supermium-portable/assets"
+	"github.com/Jeffail/gabs"
+	"github.com/pkg/errors"
+	"github.com/portapps/firefox-portable/assets"
 	"github.com/portapps/portapps/v3"
 	"github.com/portapps/portapps/v3/pkg/log"
-	"github.com/portapps/portapps/v3/pkg/registry"
+	"github.com/portapps/portapps/v3/pkg/mutex"
 	"github.com/portapps/portapps/v3/pkg/shortcut"
 	"github.com/portapps/portapps/v3/pkg/utl"
+	"github.com/portapps/portapps/v3/pkg/win"
 )
 
 type config struct {
-	Cleanup bool `yaml:"cleanup" mapstructure:"cleanup"`
+	Profile           string `yaml:"profile" mapstructure:"profile"`
+	MultipleInstances bool   `yaml:"multiple_instances" mapstructure:"multiple_instances"`
+	Locale            string `yaml:"locale" mapstructure:"locale"`
+	Cleanup           bool   `yaml:"cleanup" mapstructure:"cleanup"`
 }
 
 var (
@@ -25,47 +34,142 @@ var (
 	cfg *config
 )
 
+const (
+	defaultLocale = "en-US"
+)
+
 func init() {
 	var err error
 
 	// Default config
 	cfg = &config{
-		Cleanup: false,
+		Profile:           "default",
+		MultipleInstances: false,
+		Locale:            defaultLocale,
+		Cleanup:           false,
 	}
 
 	// Init app
-	if app, err = portapps.NewWithCfg("supermium-portable", "Supermium", cfg); err != nil {
+	if app, err = portapps.NewWithCfg("firefox-portable", "Firefox", cfg); err != nil {
 		log.Fatal().Err(err).Msg("Cannot initialize application. See log file for more info.")
 	}
 }
 
 func main() {
 	utl.CreateFolder(app.DataPath)
-	app.Process = utl.PathJoin(app.AppPath, "chrome.exe")
+	profileFolder := utl.CreateFolder(app.DataPath, "profile", cfg.Profile)
+
+	app.Process = utl.PathJoin(app.AppPath, "firefox.exe")
 	app.Args = []string{
-		"--user-data-dir=" + app.DataPath,
-		"--no-default-browser-check",
-		"--disable-logging",
-		"--disable-breakpad",
-		"--disable-machine-id",
-		"--disable-encryption-win",
+		"--profile",
+		profileFolder,
+	}
+
+	// Set env vars
+	crashreporterFolder := utl.CreateFolder(app.DataPath, "crashreporter")
+	pluginsFolder := utl.CreateFolder(app.DataPath, "plugins")
+	os.Setenv("MOZ_CRASHREPORTER", "0")
+	os.Setenv("MOZ_CRASHREPORTER_DATA_DIRECTORY", crashreporterFolder)
+	os.Setenv("MOZ_CRASHREPORTER_DISABLE", "1")
+	os.Setenv("MOZ_CRASHREPORTER_NO_REPORT", "1")
+	os.Setenv("MOZ_DATA_REPORTING", "0")
+	os.Setenv("MOZ_MAINTENANCE_SERVICE", "0")
+	os.Setenv("MOZ_PLUGIN_PATH", pluginsFolder)
+	os.Setenv("MOZ_UPDATER", "0")
+
+	// Create and check mutex
+	mu, err := mutex.Create(app.ID)
+	defer mutex.Release(mu)
+	if err != nil {
+		if !cfg.MultipleInstances {
+			log.Error().Msg("You have to enable multiple instances in your configuration if you want to launch another instance")
+			if _, err = win.MsgBox(
+				fmt.Sprintf("%s portable", app.Name),
+				"Other instance detected. You have to enable multiple instances in your configuration if you want to launch another instance.",
+				win.MsgBoxBtnOk|win.MsgBoxIconError); err != nil {
+				log.Error().Err(err).Msg("Cannot create dialog box")
+			}
+			return
+		} else {
+			log.Warn().Msg("Another instance is already running")
+		}
 	}
 
 	// Cleanup on exit
 	if cfg.Cleanup {
 		defer func() {
 			utl.Cleanup([]string{
-				path.Join(os.Getenv("APPDATA"), "Supermium"),
-				path.Join(os.Getenv("LOCALAPPDATA"), "Supermium"),
+				path.Join(os.Getenv("APPDATA"), "Mozilla", "Firefox"),
+				path.Join(os.Getenv("LOCALAPPDATA"), "Mozilla", "Firefox"),
+				path.Join(os.Getenv("USERPROFILE"), "AppData", "LocalLow", "Mozilla"),
 			})
 		}()
 	}
 
-	// Copy default shortcut
-	shortcutPath := path.Join(os.Getenv("APPDATA"), "Microsoft", "Windows", "Start Menu", "Programs", "Supermium Portable.lnk")
-	defaultShortcut, err := assets.Asset("Supermium.lnk")
+	// Locale
+	locale, err := checkLocale()
 	if err != nil {
-		log.Error().Err(err).Msg("Cannot load asset Supermium.lnk")
+		log.Error().Err(err).Msg("Cannot set locale")
+	}
+
+	// Multiple instances
+	if cfg.MultipleInstances {
+		log.Info().Msg("Multiple instances enabled")
+		app.Args = append(app.Args, "--no-remote")
+	}
+
+	// Policies
+	if err := createPolicies(); err != nil {
+		log.Fatal().Err(err).Msg("Cannot create policies")
+	}
+
+	// Autoconfig
+	prefFolder := utl.CreateFolder(app.AppPath, "defaults/pref")
+	autoconfig := utl.PathJoin(prefFolder, "autoconfig.js")
+	if err := utl.CreateFile(autoconfig, `//
+pref("general.config.filename", "portapps.cfg");
+pref("general.config.obscure_value", 0);`); err != nil {
+		log.Fatal().Err(err).Msg("Cannot write autoconfig.js")
+	}
+
+	// Mozilla cfg
+	mozillaCfgPath := utl.PathJoin(app.AppPath, "portapps.cfg")
+	mozillaCfgFile, err := os.Create(mozillaCfgPath)
+	if err != nil {
+		log.Fatal().Err(err).Msg("Cannot create portapps.cfg")
+	}
+	mozillaCfgData := struct {
+		Locale string
+	}{
+		locale,
+	}
+	mozillaCfgTpl := template.Must(template.New("mozillaCfg").Parse(`// Set locale
+pref("intl.locale.requested", "{{ .Locale }}");
+
+// Extensions scopes
+lockPref("extensions.enabledScopes", 4);
+lockPref("extensions.autoDisableScopes", 3);
+
+// Don't show 'know your rights' on first run
+pref("browser.rights.3.shown", true);
+
+// Don't show WhatsNew on first run after every update
+pref("browser.startup.homepage_override.mstone", "ignore");
+`))
+	if err := mozillaCfgTpl.Execute(mozillaCfgFile, mozillaCfgData); err != nil {
+		log.Fatal().Err(err).Msg("Cannot write portapps.cfg")
+	}
+
+	// Fix extensions path
+	if err := updateAddonStartup(profileFolder); err != nil {
+		log.Error().Err(err).Msg("Cannot fix extensions path")
+	}
+
+	// Copy default shortcut
+	shortcutPath := path.Join(os.Getenv("APPDATA"), "Microsoft", "Windows", "Start Menu", "Programs", "Firefox Portable.lnk")
+	defaultShortcut, err := assets.Asset("Firefox.lnk")
+	if err != nil {
+		log.Error().Err(err).Msg("Cannot load asset Firefox.lnk")
 	}
 	err = os.WriteFile(shortcutPath, defaultShortcut, 0644)
 	if err != nil {
@@ -77,7 +181,7 @@ func main() {
 		ShortcutPath:     shortcutPath,
 		TargetPath:       app.Process,
 		Arguments:        shortcut.Property{Clear: true},
-		Description:      shortcut.Property{Value: "Supermium Portable by Portapps and Adeii"},
+		Description:      shortcut.Property{Value: "Firefox Portable by Portapps and Adeii"},
 		IconLocation:     shortcut.Property{Value: app.Process},
 		WorkingDirectory: shortcut.Property{Value: app.AppPath},
 	})
@@ -90,41 +194,100 @@ func main() {
 		}
 	}()
 
-	// Registry keys
-	regsPath := utl.CreateFolder(app.RootPath, "reg")
-	bsRegKey := registry.Key{
-		Key:  `HKCU\SOFTWARE\Supermium`,
-		Arch: "32",
-	}
-	bbdRegKey := registry.Key{
-		Key:  `HKCU\SOFTWARE\Supermium-Browser-Development`,
-		Arch: "32",
-	}
-
-	if err := bsRegKey.Import(utl.PathJoin(regsPath, "Supermium.reg")); err != nil {
-		log.Error().Err(err).Msg("Cannot import registry key")
-	}
-	if err := bbdRegKey.Import(utl.PathJoin(regsPath, "Supermium-Browser-Development.reg")); err != nil {
-		log.Error().Err(err).Msg("Cannot import registry key")
-	}
-
-	defer func() {
-		if err := bsRegKey.Export(utl.PathJoin(regsPath, "Supermium.reg")); err != nil {
-			log.Error().Err(err).Msg("Cannot export registry key")
-		}
-		if err := bbdRegKey.Export(utl.PathJoin(regsPath, "Supermium-Browser-Development.reg")); err != nil {
-			log.Error().Err(err).Msg("Cannot export registry key")
-		}
-		if cfg.Cleanup {
-			if err := bsRegKey.Delete(true); err != nil {
-				log.Error().Err(err).Msg("Cannot remove registry key")
-			}
-			if err := bbdRegKey.Delete(true); err != nil {
-				log.Error().Err(err).Msg("Cannot remove registry key")
-			}
-		}
-	}()
-
 	defer app.Close()
 	app.Launch(os.Args[1:])
+}
+
+func createPolicies() error {
+	appFile := utl.PathJoin(utl.CreateFolder(app.AppPath, "distribution"), "policies.json")
+	dataFile := utl.PathJoin(app.DataPath, "policies.json")
+	defaultPolicies := struct {
+		Policies map[string]interface{} `json:"policies"`
+	}{
+		Policies: map[string]interface{}{
+			"DisableAppUpdate":        true,
+			"DontCheckDefaultBrowser": true,
+		},
+	}
+
+	jsonPolicies, err := gabs.Consume(defaultPolicies)
+	if err != nil {
+		return errors.Wrap(err, "Cannot consume default policies")
+	}
+	log.Debug().Msgf("Default policies: %s", jsonPolicies.String())
+
+	if utl.Exists(dataFile) {
+		rawCustomPolicies, err := os.ReadFile(dataFile)
+		if err != nil {
+			return errors.Wrap(err, "Cannot read custom policies")
+		}
+
+		jsonPolicies, err = gabs.ParseJSON(rawCustomPolicies)
+		if err != nil {
+			return errors.Wrap(err, "Cannot consume custom policies")
+		}
+		log.Debug().Msgf("Custom policies: %s", jsonPolicies.String())
+
+		jsonPolicies.Set(true, "policies", "DisableAppUpdate")
+		jsonPolicies.Set(true, "policies", "DontCheckDefaultBrowser")
+	}
+
+	log.Debug().Msgf("Applied policies: %s", jsonPolicies.String())
+	err = os.WriteFile(appFile, []byte(jsonPolicies.StringIndent("", "  ")), 0644)
+	if err != nil {
+		return errors.Wrap(err, "Cannot write policies")
+	}
+
+	return nil
+}
+
+func checkLocale() (string, error) {
+	extSourceFile := fmt.Sprintf("%s.xpi", cfg.Locale)
+	extDestFile := fmt.Sprintf("langpack-%s@firefox.mozilla.org.xpi", cfg.Locale)
+	extsFolder := utl.CreateFolder(app.AppPath, "distribution", "extensions")
+	localeXpi := utl.PathJoin(app.AppPath, "langs", extSourceFile)
+
+	// If default locale skip (already embedded)
+	if cfg.Locale == defaultLocale {
+		return cfg.Locale, nil
+	}
+
+	// Check .xpi file exists
+	if !utl.Exists(localeXpi) {
+		return defaultLocale, fmt.Errorf("XPI file does not exist in %s", localeXpi)
+	}
+
+	// Copy .xpi
+	if err := utl.CopyFile(localeXpi, utl.PathJoin(extsFolder, extDestFile)); err != nil {
+		return defaultLocale, err
+	}
+
+	return cfg.Locale, nil
+}
+
+func updateAddonStartup(profileFolder string) error {
+	lz4File := path.Join(profileFolder, "addonStartup.json.lz4")
+	if !utl.Exists(lz4File) || app.Prev.RootPath == "" {
+		return nil
+	}
+
+	lz4Raw, err := mozLz4Decompress(lz4File)
+	if err != nil {
+		return err
+	}
+
+	prevPathLin := strings.Replace(utl.FormatUnixPath(app.Prev.RootPath), ` `, `%20`, -1)
+	currPathLin := strings.Replace(utl.FormatUnixPath(app.RootPath), ` `, `%20`, -1)
+	lz4Str := strings.Replace(string(lz4Raw), prevPathLin, currPathLin, -1)
+
+	prevPathWin := strings.Replace(strings.Replace(utl.FormatWindowsPath(app.Prev.RootPath), `\`, `\\`, -1), ` `, `%20`, -1)
+	currPathWin := strings.Replace(strings.Replace(utl.FormatWindowsPath(app.RootPath), `\`, `\\`, -1), ` `, `%20`, -1)
+	lz4Str = strings.Replace(lz4Str, prevPathWin, currPathWin, -1)
+
+	lz4Enc, err := mozLz4Compress([]byte(lz4Str))
+	if err != nil {
+		return err
+	}
+
+	return os.WriteFile(lz4File, lz4Enc, 0644)
 }