+++ CHANGELOG.md
@@ -1,17 +1,293 @@
 # Changelog
 
-## 119.0.6045.192
+## 121.0-65 (2023/12/25)
 
-* Supermium 119.0.6045 (2023/12/05)
-
-## 118.0.5991.0
-
-* Supermium 118.0.5991 (2023/08/24)
+* Firefox 121.0
 * Portapps 3.9.0
 
-## 117.0.5874.0 (2023/06/xx)
+## 114.0.2-64 (2023/07/03)
 
-* Supermium 117.0.5874
+* Firefox 114.0.2
 * Portapps 3.8.0
 
-All credits goes to win32ss for making Supermium browser!
+## 109.0.1-63 (2023/02/12)
+
+* Firefox 109.0.1
+* Portapps 3.7.0
+
+## 105.0.2-62 (2022/10/04)
+
+* Firefox 105.0.2
+* Portapps 3.6.0
+* Bump github.com/kevinburke/go-bindata
+
+## 100.0.2-61 (2022/05/22)
+
+* Firefox 100.0.2
+* Portapps 3.5.0
+* Bump github.com/pierrec/lz4/v3 from 3.3.4 to 3.3.5 (#69)
+* Bump github.com/kevinburke/go-bindata (#61)
+
+## 96.0.3-60 (2022/01/27)
+
+* Firefox 96.0.3
+* Portapps 3.4.0
+
+## 92.0-59 (2021/09/19)
+
+* Firefox 92.0
+* Portapps 3.3.1
+
+## 90.0.2-58 (2021/07/25)
+
+* Firefox 90.0.2
+
+## 88.0.1-57 (2021/05/22)
+
+* Firefox 88.0.1
+
+## 88.0-56 (2021/05/22)
+
+* Firefox 88
+
+## 87.0-55 (2021/04/04)
+
+* Firefox 87
+* Portapps 3.3.0
+* No more win32 release
+
+## 85.0.1-54 (2021/02/06)
+
+* Firefox 85.0.1
+
+## 84.0.2-53 (2021/02/06)
+
+* Firefox 84.0.2
+* Portapps 3.1.0
+
+## 83.0-52 (2020/12/07)
+
+* Firefox 83.0
+* Portapps 3.0.3
+
+## 82.0.2-51 (2020/10/28)
+
+* Firefox 82.0.2
+
+## 81.0.1-50 (2020/10/09)
+
+* Firefox 81.0.1
+
+## 81.0-49 (2020/09/29)
+
+* Firefox 81.0
+
+## 80.0.1-48 (2020/09/14)
+
+* Firefox 80.0.1
+
+## 80.0-47 (2020/08/24)
+
+* Firefox 80.0
+* Portapps 2.6.0
+
+## 79.0-47 (2020/07/31)
+
+* Firefox 79.0
+
+## 78.0.2-46 (2020/07/20)
+
+* Firefox 78.0.2
+
+## 78.0.1-45 (2020/07/08)
+
+* Firefox 78.0.1
+
+## 78.0-44 (2020/06/29)
+
+* Firefox 78.0
+
+## 77.0.1-43 (2020/06/13)
+
+* Firefox 77.0.1
+
+## 76.0.1-42 (2020/05/29)
+
+* Fix addons startup (#3) 
+* Portapps 2.4.4
+
+## 76.0.1-41 (2020/05/10)
+
+* Firefox 76.0.1
+* Allow custom `policies.json` (#25)
+* `DisableTelemetry` and `DisableFirefoxStudies` removed (use policies file instead)
+* Portapps 2.2.4
+
+## 75.0-40 (2020/04/12)
+
+* Firefox 75.0
+* Portapps 2.0.5
+
+## 74.0.1-39 (2020/04/12)
+
+* Firefox 74.0.1
+* Portapps 2.0.2
+
+## 74.0-38 (2020/03/12)
+
+* Firefox 74.0
+
+## 73.0.1-37 (2020/02/24)
+
+* Firefox 73.0.1
+
+## 73.0-36 (2020/02/12)
+
+* Firefox 73.0
+
+## 72.0.2-35 (2020/01/25)
+
+* Firefox 72.0.2
+
+## 72.0.1-34 (2020/01/09)
+
+* Firefox 72.0.1
+
+## 72.0-33 (2020/01/06)
+
+* Firefox 72.0
+* Add `cleanup` config
+* Portapps 1.31.0
+
+## 71.0-32 (2019/12/03)
+
+* Firefox 71.0
+
+## 70.0.1-31 (2019/12/01)
+
+* Following a trademark violation report (#11), name and logo have been changed
+
+## 70.0.1-30 (2019/11/03)
+
+* Firefox 70.0.1
+
+## 70.0-29 (2019/10/29)
+
+* Update Firefox logo
+
+## 70.0-28 (2019/10/22)
+
+* Firefox 70.0
+
+## 69.0.3-27 (2019/10/11)
+
+* Firefox 69.0.3
+* Portapps 1.28.0
+
+## 69.0.2-26 (2019/10/06)
+
+* Firefox 69.0.2
+* Portapps 1.27.0
+
+## 69.0.1-25 (2019/09/18)
+
+* Firefox 69.0.1
+
+## 69.0-24 (2019/09/09)
+
+* Fix addons disabled on startup (#3)
+* Portapps 1.26.1
+
+## 69.0-23 (2019/09/02)
+
+* Firefox 69.0
+* Portapps 1.26.0
+
+## 68.0.2-22 (2019/08/14)
+
+* Firefox 68.0.2
+* Portapps 1.25.0
+
+## 68.0.1-21 (2019/07/19)
+
+* Firefox 68.0.1
+
+## 68.0-20 (2019/07/08)
+
+* Firefox 68.0
+
+## 67.0.4-18 (2019/06/21)
+
+* Firefox 67.0.4
+
+## 67.0.3-17 (2019/06/18)
+
+* Firefox 67.0.3
+
+## 67.0.2-16 (2019/06/10)
+
+* Firefox 67.0.2
+
+## 67.0.1-15 (2019/06/06)
+
+* Extensions break on path change (#2)
+
+## 67.0.1-13 (2019/06/04)
+
+* Firefox 67.0.1
+
+## 67.0-12 (2019/05/21)
+
+* Add config param to change locale
+* Make Firefox multilingual
+
+## 67.0-11 (2019/05/20)
+
+* Firefox 67.0
+* Portapps 1.24.1
+
+## 66.0.5-10 (2019/05/08)
+
+* Firefox 66.0.5
+* Portapps 1.23.0
+
+## 66.0.4-9 (2019/05/05)
+
+* Firefox 66.0.4
+* Portapps 1.22.2
+
+## 66.0.3-8 (2019/04/10)
+
+* Firefox 66.0.3
+
+## 66.0.2-7 (2019/03/27)
+
+* Firefox 66.0.2
+* Portapps 1.21.0
+
+## 66.0.1-6 (2019/03/23)
+
+* Firefox 66.0.1
+
+## 66.0-5 (2019/03/20)
+
+* Firefox 66.0
+
+## 65.0.2-4 (2019/03/18)
+
+* Handle notifications
+* Portapps 1.20.3
+
+## 65.0.2-3 (2019/03/06)
+
+* Use [latest logo design](https://design.firefox.com/photon/visuals/product-identity-assets.html)
+
+## 65.0.2-2 (2019/03/06)
+
+* Fix configuration values not handled
+* Display error box if Firefox is already running and multiple instances disabled
+* Create profile folder at first launch
+
+## 65.0.2-1 (2019/03/05)
+
+* Initial version based on Firefox 65.0.2